Finalized Draft for Player Movement + Dash Mechanic
The following was completed under my player movement scene: Utilized scripting to complete draft for player movement using WASD keys. Allows for movement in eight directions under an equal changeable speed. For future options and potential variance in gameplay, implemented "dash" mechanic which changes movement speed for a custom amount of time while also having a custom cooldown time. Whether or not it be included in the final build is a topic of discussion.